add precipitation data grid and importer
This model functions similarly to the weather data importer (another NOAA gridded product). Instead of trying to shoehorn the precip data into the existing weather model I created a new model and importer. Need to create a precip controller to allow API access, and add precip to the Soils and WISP sites.